+++ src/main/java/com/board/controller/BoardController.java
@@ -1,9 +1,9 @@
 package com.board.controller;
 
-import com.board.controller.dto.BoardDto;
-import com.board.controller.dto.CreateBoardRequestDto;
-import com.board.controller.dto.ListBoardResponseDto;
-import com.board.controller.dto.UpdateBoardRequestDto;
+import com.board.controller.dto.board.BoardDto;
+import com.board.controller.dto.board.CreateBoardRequestDto;
+import com.board.controller.dto.board.ListBoardResponseDto;
+import com.board.controller.dto.board.UpdateBoardRequestDto;
 import com.board.service.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -17,7 +17,7 @@ public class BoardController {
     private final BoardService boardService;
 
     @PostMapping
-    public ResponseEntity createBoard(@RequestBody CreateBoardRequestDto createBoardRequestDto) {
+    public ResponseEntity<?> createBoard(@RequestBody CreateBoardRequestDto createBoardRequestDto) {
         boardService.createBoard(createBoardRequestDto);
 
         return ResponseEntity.ok().build();
@@ -36,13 +36,13 @@ public class BoardController {
     }
 
     @DeleteMapping(path = "/{boardId}")
-    public ResponseEntity deleteBoard(@PathVariable(name = "boardId") Long boardId) {
+    public ResponseEntity<?> deleteBoard(@PathVariable(name = "boardId") Long boardId) {
         boardService.deleteBoard(boardId);
         return ResponseEntity.ok().build();
     }
 
     @PatchMapping("/{boardId}")
-    public ResponseEntity<UpdateBoardRequestDto> updateBoard(@PathVariable(name = "boardId") Long boardId, @RequestBody UpdateBoardRequestDto updateBoardRequestDto) {
+    public ResponseEntity<?> updateBoard(@PathVariable(name = "boardId") Long boardId, @RequestBody UpdateBoardRequestDto updateBoardRequestDto) {
         boardService.updateBoard(boardId, updateBoardRequestDto);
 
         return ResponseEntity.ok().build();
