+++ src/main/java/com/board/controller/ReplyController.java
@@ -1,9 +1,9 @@
 package com.board.controller;
 
-import com.board.controller.dto.CreateReplyRequestDto;
-import com.board.controller.dto.ListReplyResponseDto;
-import com.board.controller.dto.ReplyDto;
-import com.board.controller.dto.UpdateReplyRequestDto;
+import com.board.controller.dto.reply.CreateReplyRequestDto;
+import com.board.controller.dto.reply.ListReplyResponseDto;
+import com.board.controller.dto.reply.ReplyDto;
+import com.board.controller.dto.reply.UpdateReplyRequestDto;
 import com.board.service.ReplyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +19,7 @@ public class ReplyController {
     private final ReplyService replyService;
 
     @PostMapping(path = "/{commentId}/reply")
-    public ResponseEntity createReply(@PathVariable(name = "commentId") Long commentId, @RequestBody CreateReplyRequestDto createReplyRequestDto) {
+    public ResponseEntity<?> createReply(@PathVariable(name = "commentId") Long commentId, @RequestBody CreateReplyRequestDto createReplyRequestDto) {
         replyService.createReply(commentId, createReplyRequestDto);
 
         return ResponseEntity.ok().build();
@@ -43,13 +43,13 @@ public class ReplyController {
     }
 
     @PatchMapping(path = "/reply/{replyId}")
-    public ResponseEntity updateReply(@PathVariable(name = "replyId") Long replyId, @RequestBody UpdateReplyRequestDto updateReplyRequestDto) {
+    public ResponseEntity<?> updateReply(@PathVariable(name = "replyId") Long replyId, @RequestBody UpdateReplyRequestDto updateReplyRequestDto) {
         replyService.updateReply(replyId, updateReplyRequestDto.getContent());
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping(path = "/reply/{replyId}")
-    public ResponseEntity deleteReply(@PathVariable(name = "replyId") Long replyId) {
+    public ResponseEntity<?> deleteReply(@PathVariable(name = "replyId") Long replyId) {
         replyService.deleteReply(replyId);
         return ResponseEntity.ok().build();
     }
