--- src/main/java/com/board/controller/dto/GetCommentResponse.java
@@ -1,11 +1,0 @@
-package com.board.controller.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class GetCommentResponse {
-
-    private final GetCommentResponseDto getComment;
-}
