+++ src/main/java/com/board/controller/dto/board/BoardDto.java
@@ -1,4 +1,4 @@
-package com.board.controller.dto;
+package com.board.controller.dto.board;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
