+++ src/main/java/com/board/controller/dto/board/CreateBoardRequestDto.java
@@ -1,12 +1,13 @@
-package com.board.controller.dto;
+package com.board.controller.dto.board;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class CreateCommentRequestDto {
+public class CreateBoardRequestDto {
 
+    private String title;
     private String content;
     private String userId;
