+++ src/main/java/com/board/controller/dto/board/UpdateBoardRequestDto.java
@@ -1,12 +1,13 @@
-package com.board.controller.dto;
+package com.board.controller.dto.board;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class CreateReplyRequestDto {
+public class UpdateBoardRequestDto {
 
+    private String title;
     private String content;
-    private String userId;
+
 }
