+++ src/main/java/com/board/controller/dto/comment/CreateCommentRequestDto.java
@@ -1,13 +1,12 @@
-package com.board.controller.dto;
+package com.board.controller.dto.comment;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class CreateBoardRequestDto {
+public class CreateCommentRequestDto {
 
-    private String title;
     private String content;
     private String userId;
