+++ src/main/java/com/board/controller/dto/comment/UpdateCommentRequestDto.java
@@ -1,4 +1,4 @@
-package com.board.controller.dto;
+package com.board.controller.dto.comment;
 
 import lombok.*;
