+++ src/main/java/com/board/controller/dto/reply/CreateReplyRequestDto.java
@@ -1,13 +1,12 @@
-package com.board.controller.dto;
+package com.board.controller.dto.reply;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class UpdateBoardRequestDto {
+public class CreateReplyRequestDto {
 
-    private String title;
     private String content;
-
+    private String userId;
 }
