+++ src/main/java/com/board/controller/dto/reply/ReplyDto.java
@@ -1,7 +1,6 @@
-package com.board.controller.dto;
+package com.board.controller.dto.reply;
 
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
