+++ src/main/java/com/board/repository/ReplyRepository.java
@@ -1,6 +1,5 @@
 package com.board.repository;
 
-import com.board.controller.dto.ReplyDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
